codes/pid_controller/302_Kodu/requirements.txt
numpy==1.12.1
pyqtgraph==0.10.0
pyserial==3.4